added band limited sawtooth

+++ notebuf/buffer.py
@@ -10,7 +10,7 @@ class Buffer(_Param):
         super().__init__(params)
 
         if not isinstance(buff, np.ndarray):
-            self.buff = np.arange(int(self.sample_rate * self.duration), dtype=np.float32)
+            self.buff = np.arange(int(self.sample_rate * self.duration), dtype=np.float64)
         else:
             self.buff = buff
